data/requirements.txt
faiss-cpu==1.8.0
protobuf==4.25.3
streamlit==1.36.0
langchain==0.3.24
langchain-community==0.3.5
langchain-huggingface==0.1.2
python-dotenv